create prover module
