+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod datetime_utils;
 pub mod proof_type;
+pub mod prover;
 pub mod task_state;
 pub mod types;
